app/pyproject.toml
[tool.poetry]
name = "squidbot"
version = "0.1.0"
description = "Autonomous AI agent with Telegram, tool chaining, skills, and proactive messaging"
authors = ["leejw51"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/leejw51/SquidBot"
repository = "https://github.com/leejw51/SquidBot"
keywords = ["ai", "agent", "telegram", "bot", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "squidbot"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.0"
python-telegram-bot = "^20.0"
playwright = "^1.40"
duckduckgo-search = "^6.0"
apscheduler = "^3.10"
python-dotenv = "^1.0"
aiofiles = "^23.0"
aiosqlite = "^0.19"
numpy = "^1.26"
# Web3 plugin dependencies
web3 = "^7.0"
mnemonic = "^0.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"
pytest-cov = "^4.0"

[tool.poetry.scripts]
squidbot = "squidbot.server:main"
squidbot-client = "squidbot.client:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
